# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CineTrack API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the CineTrack Fastify backend."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "List movies",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rating_usia",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "movie_id": "M003",
                      "title": "Agak Laen",
                      "genre": [
                        "Comedy"
                      ],
                      "rating_usia": "R13",
                      "duration_min": 132
                    }
                  ],
                  "meta": {
                    "filters": {
                      "search": null,
                      "genre": null,
                      "rating_usia": null
                    },
                    "pagination": {
                      "page": 1,
                      "limit": 20,
                      "total": 12,
                      "total_pages": 1
                    }
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/studios": {
      "get": {
        "tags": [
          "Studios"
        ],
        "summary": "List studios",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "screen_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Studio list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "studio_id": "ST-C001-1",
                      "cinema_id": "C001",
                      "studio_name": "Studio 1",
                      "total_capacity": 150,
                      "screen_type": 2
                    }
                  ],
                  "meta": {
                    "filters": {
                      "cinema_id": null,
                      "studio_id": null,
                      "screen_type": null
                    },
                    "pagination": {
                      "page": 1,
                      "limit": 20,
                      "total": 60,
                      "total_pages": 3
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/studios/{studio_id}": {
      "get": {
        "tags": [
          "Studios"
        ],
        "summary": "Get studio detail",
        "parameters": [
          {
            "$ref": "#/components/parameters/StudioIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Studio detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "studio_id": "ST-C001-1",
                    "cinema_id": "C001",
                    "studio_name": "Studio 1",
                    "total_capacity": 150,
                    "screen_type": 2,
                    "total_schedules": 360,
                    "total_tickets": 15116
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/schedules": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "List schedules",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "movie_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "show_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "schedule_id": "S021428",
                      "movie_id": "M008",
                      "studio_id": "ST-C006-2",
                      "cinema_id": "C006",
                      "show_date": "2026-05-14",
                      "start_time": "21:00",
                      "price": 35000,
                      "status": "Available"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/schedules/{schedule_id}": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "Get schedule detail",
        "parameters": [
          {
            "$ref": "#/components/parameters/ScheduleIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "schedule_id": "S021428",
                    "movie_id": "M008",
                    "studio_id": "ST-C006-2",
                    "cinema_id": "C006",
                    "show_date": "2026-05-14",
                    "start_time": "21:00",
                    "price": 35000,
                    "status": "Available",
                    "tickets_sold": 19,
                    "revenue": 770000
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/tikets": {
      "get": {
        "tags": [
          "Tikets"
        ],
        "summary": "List tikets",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "schedule_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "payment_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seat_category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Tiket list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "tiket_id": "T00911997",
                      "schedule_id": "S021576",
                      "movie_id": "M011",
                      "cinema_id": "C018",
                      "seat_category": "Regular",
                      "final_price": 35000,
                      "trans_time": "2026-05-14T19:00:00Z",
                      "payment_type": "DEBIT"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/tikets/{tiket_id}": {
      "get": {
        "tags": [
          "Tikets"
        ],
        "summary": "Get tiket detail",
        "parameters": [
          {
            "$ref": "#/components/parameters/TiketIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Tiket detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "tiket_id": "T00911997",
                    "schedule_id": "S021576",
                    "movie_id": "M011",
                    "studio_id": "ST-C018-3",
                    "cinema_id": "C018",
                    "seat_category": "Regular",
                    "final_price": 35000,
                    "trans_time": "2026-05-14T19:00:00Z",
                    "payment_type": "DEBIT"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "List tickets",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "schedule_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "payment_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seat_category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "ticket_id": "T00000001",
                      "schedule_id": "S000001",
                      "movie_id": "M001",
                      "cinema_id": "C001",
                      "seat_category": "Regular",
                      "final_price": 45000,
                      "trans_time": "2026-05-14T12:00:00.000Z",
                      "payment_type": "QRIS",
                      "status": "success",
                      "is_mock_status": true
                    }
                  ],
                  "meta": {
                    "filters": {
                      "schedule_id": null,
                      "movie_id": null,
                      "cinema_id": null
                    },
                    "pagination": {
                      "page": 1,
                      "limit": 20,
                      "total": 1,
                      "total_pages": 1
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Get ticket detail",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "ticket_id": "T00000001",
                    "schedule_id": "S000001",
                    "movie_id": "M001",
                    "cinema_id": "C001",
                    "seat_category": "Regular",
                    "final_price": 45000,
                    "trans_time": "2026-05-14T12:00:00.000Z",
                    "payment_type": "QRIS",
                    "status": "success",
                    "is_mock_status": true
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found"
              }
            }
          }
        }
      }
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "required": false
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "required": false
      },
      "StartDateParam": {
        "name": "start_date",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date"
        },
        "required": false
      },
      "EndDateParam": {
        "name": "end_date",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date"
        },
        "required": false
      },
...
      "CinemaIdQueryParam": {
        "name": "cinema_id",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "required": false
      },
...
      "StudioIdParam": {
        "name": "studio_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "ScheduleIdParam": {
        "name": "schedule_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "TiketIdParam": {
        "name": "tiket_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {},
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": [
                  "array",
                  "object",
                  "null"
                ],
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
...
    }
  }
}